+++ inc/config.h
@@ -37,6 +37,7 @@ struct game_config {
 };
 
 gboolean parse_ini_file(const char *filename, struct game_config *config);
+void write_ini_file(const char *filename);
 
 /* parse the command line */
 void parse_commandline(int argc, char *argv[], struct game_config *config);